更新配置文件，更改组件，支持JSX
[Chore]
更新vite.config.ts配置。

[Fix]
移除header文件夹下的部分vue组件，Dark组件现在是DarkSwitch.tsx。
