--- src/renderer/src/components/header/Title.vue
@@ -1,6 +1,0 @@
-<template>
-    <div class="flex flex-col flex-none">
-        <h1 class="text-base">To-Do</h1>
-        <p class="text-xs">Preview</p>
-    </div>
-</template>
